Fix sector move speed / optimize animation ticks

+++ clientd3d/room.h
@@ -55,6 +55,6 @@ typedef struct
    DWORD overrideDepth[4];
 } room_type;
 
-Bool AnimateRoom(room_type *room, int dt);
+Bool AnimateRoom(room_type *room, float dt);
 
 #endif /* #ifndef _ROOM_H */
